Change RUN for ensemble jobs
Updates `$RUN` and `$CDUMP` for ensemble jobs to include `enkf`.

Previously, the `$RUN` for EnKF jobs had been set to `gdas` or `gfs`.
However, this violates NCO policy and was also complicating Issue #761.
Now the `$RUN` for EnKF jobs is either `enkfgdas` or `enkfgfs`.

Theoeretically, `$CDUMP` shouldn't need to change. However, `$CDUMP`
and `$RUN` are used interchangably throughout much of the workflow
(Issue #1299), so for now the `$CDUMP` is kept identical to `$RUN`.

Refs: #1298