GFSv16.3.6 (#1304)
RFC 10562 - On WCOSS2, implement a modified GFS global analysis to remove
the use of GMI (Global Precipitation Measurement Microwave Imager) data. The
developer noticed an array overflow in the processing of GMI data on Hera
which may lead to a SEGFAULT for the GSI (Grid-Point Statistical 
Interpolation). Removing this data usage will prevent a potential operational
GSI crash.

Refs #1278