Update rocoto data dependencies to use templates
Updates the rocoto data dependencies to determine their paths based
on the COM templates found in `config.com`. To facilitate this, a
function is added to `workflow_tasks.py` that takes in a string and
makes replacements suitable for use in a rocoto <cyclestr> tag. This
function leverages pygw.template.

Refs: #761