Move generate_com to preamble and improve
The `generate_com` function is moved out of `config.com` and into
the preamble. This is a temporary home until a new common functions
script is created. This move is done for a couple reasons:
* Moves the function out of a config file
* Mitigates a bug in bash < 4.4 that breaks `$BASH_SOURCE` (and thus
  the trace message) when using functions exported by a parent script

The `generate_com` function is also improved to allow the definition
of a template to use, overriding the default of `${var}_TMPL`. This
should allow the useage of `generate_com` in all instances instead of
needing to use `envsubst` directly in scripts for alternative cases
(like when paths are needed for multiple times). Since the function
allows for multiple variables at once, the defining of template name
is done within a single argument by separating the variable and the
template name with a colon. Without a colon, the argument behaves the
same as previously, with the template `${var}_TMPL` being used.

The trace string is also updated to more elegantly handle tracing
through functions and unknown sources.

Refs: #761