Complete update to arch jobs COM refactor
Completes updates of the archive jobs to use the COM templates. The
gdas section hadn't been properly updated, and the coupled components
also needed to be completed.

Also updated so that the tarballs only contain relative paths as they
did before instead of absolute paths. This is done by removing `$ROTDIR`
from the COM variables when constructing the lists of files to tar.

Refs: #761