Update gempak jobs for COM refactor
Updates the gempak jobs to use the new COM templates. Best effort
was made to update the gdas versions, but those are untested as
gdas gempack is not currently available in development mode.

The end time for the plots is modified to be `$FHMAX_GFS` rather
than a hard-coded value to allow the scripts to complete successfully
when run for less than the operational length.

Refs: #761