Update AWIPS for COM refactor
AWIPS jobs are updated to use the new COM templates. The job currently
completes up until the final program call (`mkgfsawps.x`), which errors
for a yet-unknown reason.

An additional step was added remove `5WAVH` from the expected parameter
list when the wave model is not run. This allows the job to run when
running with waves off in development.

Refs: #761