Update exp setup to stage ICs in new COM structure
Updates `setup_expt.py` to stage initial conditions in the new
hierarchical COM structure. For flexibility, the source directory
can either be in the new structure or the old flatter structure.

These paths are hard-coded, so if there are updates to any of the
restart, analysis, or input templates, those changes need to be
mirrored here.

Refs: #761