Use longer of $ILPOST and $finc for short-range coarse res gempak
The atmos gempak job would look for post output at a frequency of `$finc`
for non-0p25 resolutions before 240 h. However, if the output frequency
was shorter than `$finc`, this would result in checking for files that
aren't produced. Now the frequency will be the larger of `$ILPOST` and
`$finc`.

Refs: #761