Move input.nml to history directory
`input.nml` was formerlly being placed in the `COM_ATMOS_INPUT` directory.
However, this directory is only present for cold-starts. This led to the
file being linked as the directory name because the filename wasn't included
being used for the link target. While we could create an input directory to
hold just the one file, that is unnecessary and is likely not the best
location for the file anyway. Now it is placed in the history directory.

Refs: #761