Change wave buoy file to symlink
The wave buoy file used was simply a copy of the wave_gfs.buoys.dat file.
The .dat file is a truncated form of the wave_gfs.buoys.full file that
includes no boundary points, which prevents the testing of wave boundary
jobs. The buoys file is changed from a copy of the .dat file to a symlink
that points to the full file.

Any user wishing to use the truncated file can change the target of the
wave_gfs.buoys symlink.

Fixes: #1464