Fix wave boundary point jobs
Some minor updates were needed to get the boundary point jobs working.
- `FHMAX_WAV_IBP` had been set in the bndpnt config file but not used
in the j-job. This was invisible unless a user changed the value, since
the config and j-job used the same default.
- Checks against `FHMAX_WAV` would set the unused `FHMAX_WAV_IBP` to
the max value instead of the `FHMAX_WAV_PNT` used for the loop. This is
a problem when running for less than 180 h (the default value).
- The boundary point bulletin job was not in the env job list for Orion.

Now the boundary point jobs set `FHMAX_WAV_PNT` to `$FHMAX_WAV_IBP`.
`FHMAX_WAV_IBP` was moved from the bndpnt config to config.wave so it is
visible to bndpntbll as well.

Refs: #1464