Split MPMD stdout into tasks on slurm (#1469)
It can be difficult to debug MPMD jobs because their logs are all written concurrently to a single file. While the use of tags to designate which task via the preamble and PS4 can help identify which line is from which task, it is still difficult to follow a single task through the log, particularly for larger MPMD jobs with dozens of tasks.

Individual stdout files are now created by using the `srun` `--output` option. These files are written to the working directory (in `$DATA`).

Fixes: #1468