Move ESMF summary to atmos history dir
The ESMF timing summary had been put in `$COM_TOP`, but that doesn't
work with ensemble members, as they would overwrite each other. The
file is moved to the atmos history directory for want of a better
space, since atmos always runs and I'm unwilling to make a new directory
for a single diagnostic file that isn't always generated.

Refs: #761