Refactor UFS-DA ATM var component to use python g-w (#1372)
This PR contains UFS-DA ATM var changes in PR https://github.com/NOAA-EMC/global-workflow/pull/1354.   

Below is a list of changes in this PR
- rename UFS-DA ATM var jobs `atmanalprep`, `atmanalrun`, and `atmanalpost` as `atmanlinit`, `atmanlrun`, and `atmanlfinal`, respectively
- replace UFS-DA ATM var shell scripts with python scripts
- rename UFS-DA ATM var j-jobs consistent with initialize, run, and finalize functions.  Update j-jobs to execute python scripts instead of shell scripts
- rename UFS-DA ATM var rocoto jobs to be consistent with initialize, run, and finalize functions.  Update jobs to set python paths and execute renamed j-jobs
- update rocoto workflow generation to new names for UFS-DA ATM var jobs 
- update UFS-DA ATM var job names in machine dependent `env` files to new job names
- rename UFS-DA ATM var configuration files consistent with change in job names
- add UFS-DA ATM yaml files for UFS-DA fix files, crtm files, and increments
- add python class for UFS-DA ATM var analysis
- link UFS-DA python increment conversion script from GDASApp to g-w `ush` directory

The above changes are part of a larger g-w effort to transition from shell scripts to python.  UFS-DA Aerosol was the first GDASApp system to be converted.   This PR represents the second GDASApp system, UFS-DA atmospheric variational DA, to be converted.  PR #1373 contains changes to convert the UFS-DA ensemble DA to a python based approach.   

Fixes (in part) #1313
Depends (in part) on PR #1370 and [GDASApp #388](https://github.com/NOAA-EMC/GDASApp/pull/388)