Add GEFS capability to setup_expt (#1582)
Allows GEFS experiment directories to be set up using `setup_expt.py`. This
only creates the experiment directory; workflow generation will be added 
in a follow-up PR.

To facilitate this, a new positional argument is added in the first position to
specify which system to set up: `gfs` or `gefs`. If gfs is chosen, the second
positional argument is to choose the mode, and allows the same values as
before: `cycled` and `forecast-only`. If gefs is chosen, forecast-only is
assumed and no second positional argument is required.

To allow both gfs and gefs configure files to exist in workflow concurrently,
the `parm/config` directory contains two new directories: gfs and gefs. All
of the old config files have been moved into the gfs directory. A few select
config files have also been copied to the gefs directory. As yet, they are un-
modified (except the changes noted below). These config files will be
updated in a future PR so the modifications for GEFS are not hidden in the
new file creation. The `yaml/defaults` is also copied over, but the contents
were removed since GEFS will not be running its own analyses.

A couple variables have been renamed to be more appropriate for use in
both systems: `NMEM_ENKF`, `NMEM_EFCS`, and `CASE_ENKF` are now
`NMEM_ENS`, `NMEM_ENS_GFS`, and `CASE_ENS` respectively. Further,
`HOMEfv3gfs` were removed as unused.

Finally, resolution validation was extended to the ensemble resolution as well.

Fixes: #1576