Avoid parsing group name when checking RadMon diagnostic files (#1559)
This changes the way the RadMon diagnostic file checker determines if a diagnostic
file is 0-sized.  Instead of attempting to parse the output of `tar -tv`, the script now
extracts the `radstat` file and runs `du -b` on each gzipped diagnostic file.  On S4, the
primary group is `domain users`, which, containing a space, caused issues with the
previous method.

Fixes #1515