Update config.base to define waveGRD based on ATM res
config.base is updated to adjust the `waveGRD` used based on the
atmosphere resolution rather than one value for all resolutions.
Like the similar setting of ocean resolution, this is temporary
until the workflow is updated to allow arbitrary combinations.

Pairings of waveGRD to resolution are provisional and subject to
review by Jessica.