Fix daily file location in archive job
During the COM refactor, a location for the ocn/ice daily files was
defined within the products directory. However, the daily files are
output directly by the model and land in the history directory. The
archive job was looking in the otherwise unused daily product
directory for the files and not finding them. This update now:
- Removes the COM variable for the daily files
- Updates the archive job to pull the daily files from history

Fixes #1709