Fix ocean file locations in archive job (#1716)
During the COM refactor, a location for the ocn/ice daily files was defined within the products directory. However, the daily files are output directly by the model and land in the history directory. The archive job was looking in the otherwise unused daily product directory for the files and not finding them.

The 2D and 3D ocean files had the opposite problem: archive was looking in the history directory, but these files have their own directories in products.

This update now:
- Removes the COM variable for the daily files
- Updates the archive job to:
  - pull the daily ocean files from history
  - pull 2D ocean files from the 2D product directory
  - pull 3D ocean files from the 3D product directory
 
Fixes #1709