Add S2SA nems.configure template (#1593)
* Add S2SA nems.configure template

Most of the configuration was setup to allow S2SA, but there was no
appropriate nems.configure template. A new one was created by
starting with the S2SWA template and removing the wave portions.

Also added two UFSDA scripts that were missing from the git ignore
list.

Fixes #1592

* Update documentation for cycled mode setup

The experiment setup documentation for cycled mode had not been
updated as coupled capability was added. Also added notes to both
the cycled and forecast-only section that ATMW is currently not
working.

* Update setup_expt to recognize more S2S cycled exps

setup_expt would only detect two of the possible S2S options when
staging cycled ICs.

Refs: #1592

* Change path to volcanic aerosol data

The volcanic aerosol data was moved to a new location within the
GOCART emissions directory.

Refs: #1592

* Disable enkf versions of coupled settings

The enkf versions of coupled settings (DO_OCN, etc.) do not work
properly because some settings have already been determined by that
point based on the non-enkf versions of the settings. The enkf
versions are now disabled until there is a solution (see #1692).

Refs: #1593
Refs: #1692

* Fix ocnpost detection for gfs free-forecast

The task list for the free-forecast was checking against a list of
apps that did not include S2SA to determine whether ocnpost is
included. That check is now changed to check on do_ocean rather than
looking at the app directly.

Refs: #1593