Support coupled GEFS forecast, use mem000 for GEFS control
Makes the necessary updates to allow the workflow to run (non-wave) coupled
forecasts. Some of the coupled functions in postdet had to be updated because
they only checked for gfs and gdas runs, so necessary code was never executed.

Also updates so that the GEFS control uses `mem000` as the memdir so it
appears at the same level as the perturbed members. Relatedly, the default
member number in coupled files is changed to 000 instead of 01.

Finally, an unrelated error in the earc job is corrected where a character
from the middle of the line was moved to the beginning.

Resolves #1754