Fix SDATE_GFS issue in cycled and forecast-only modes (#1753)
On further examination, `SDATE_GFS` in the context of `metp` task is the first date that METP was run and started generating statistics.  This is different when doing cycled (`gfs_cyc` determines `SDATE_GFS`).  In the case of forecast-only, `SDATE_GFS` is always the date the first free-forecast was launched (`SDATE`)

While the fix by @RussTreadon-NOAA worked for cycling, it broke for forecast-only setups and needed to account for the other mode of operation.

Fixes #1748