Prepare GTS snow depth observations for JEDI-based Land DA (#1761)
This PR:

- adds a job to prepare GTS snow depth observations as a task in the workflow.
This task depends on the `prep.sh` job to bring GTS adpsfc bufr data. To test this
type of data DMPDIR in `config.base` needed to be pointed to "/scratch1/NCEPDEV/global/Jiarui.Dong/JEDI/GlobalWorkflow/para_gfs/glopara_dump". 
- land_analysis.py introduces a method `prepare_GTS` for this type of data.
- Updates are necessary in the GDASApp repo. See companion PR NOAA-EMC/GDASApp#541

The `preplandobs` job runs at the all four cycles in the workflow. The `prepare_IMS` job runs
only at 18z cycle, and this is controlled in the script.