Add soil color data and change the way to read in Noah-MP parameter table (#1697)
- A new soil color dataset is introduced to Noah MP to reflect the right soil albedo to reduce
the large warm bias found in the Sahel desert. The more-recent and higher resolution BNU
soil texture data will be used to replace the old data. 

- Noah MP parameter table is part of the source code in the current version. The UFS model
has to be recompiled whenever there is any parameter change in the table. Therefore we
change this structure and the table will become one of the input file to the model.

Closes #1614 
Closes #1617