Move GDASApp to CRTM 2.4 (#1773)
This PR moves the GDASApp to CRTM 2.4. It also removes the need for the global-workflow
team to maintain CRTM coefficients in their fix space, instead we should be linking/copying
them from wherever the CRTM coefficients are stored (like the way GSI does it). For now,
that is an 'unofficial' location, but will eventually be the 'official' one.