Convert v16.3.8 to satingest update instead of obsproc (#1804)
A temporary fix to allow the satingest changes to be implemented on August 29 without causing GSI to crash. Issue with the CRTM means that GSI will crash unless NOAA-21 radiances are switched off.

Obsproc version numbers should remain unchanged unless the obsproc changes are implemented before or at the same time as the satingest changes.