Allow sorc/build_tropcy_NEMS.sh to work in Release/gfs.v16.3.8 (#1815)
For some reason the sorc/build_tropcy_NEMS.sh no longer works on WCOSS2 machines.   @RussTreadon-NOAA and @aerorahul found fixes that allow the build to complete.  This PR implements one of these fixes - removing the references to the NEMSIOGFS module, which is being loaded but does not have a version number attached.

Six executables result from this build as expected.   They should still be tested for functionality.

This references #1812.