Final Updates for HR2 (#1827)
This PR has the final updates for HR2.  There was one added variable for diagnostic output: 
`iopt_diag=3` and required from land team. This also points to HR2 updates for initial conditions
which update the land states in `sfc*` files compared to HR1. Lastly, it was determined that the
setting to enable running on hera `WRTTASK_PER_GROUP_PER_THREAD_PER_TILE_GFS=20`
actually will fail on WCOSS2. While not ideal, a setting that would run out of the box on both
hera and wcoss2 for C768 could not be found, therefore a comment was added noting WCOSS2
needs to use a different setting until it can be resolved.

Resolves #1500