Simplify downstream atmos product generation scripts (#1822)
This PR came about attempting to understand how the atmos downstream products are
generated from the master grib2 files before they are updated to python-based scripts
and improved control.

This PR:
- removes explicit if blocks over the two downsets that largely repeats the same
operations and condenses into a loop
- introduces a generic `run_mpmd.sh` that is tested on Hera and WCOSS2 and can be
used elsewhere in the global-workflow.  Its input is a script that needs the instructions
to be executed, one-per-line.

No change in information content is expected from this update to the products.

This PR is necessary if the global-workflow team is to mange the post processing and
product generation scripts.