Meld S2SW config back into regular files
When the prototype S2S was first added, we placed some settings in
a `config.defaults.s2sw` file to avoid interference with existing
settings. Now that we are moving towards v17 and run S2S regularly,
it is time to merge these settings back into the other config files.

The wave output frequency settings are updated to the coarser fre-
quency as that is more suitable for development. The prior (ops)
settings are noted as comments.

There are a couple of settings that control ice fields within FV3.
These are placed in `config.ice`, so when the ice model is on those
settings are still used.

The forecast length overrides are just dropped as the existing
ones have already been shortened (though not as much) to be suitable
for most development.

"Gaussian" archive is updated to the former override values.

Resolves #1606