Merge S2S defaults config back into other configs (#1876)
When the prototype S2S was first added, we placed some settings in
a `config.defaults.s2sw` file to avoid interference with existing
settings. Now that we are moving towards v17 and run S2S regularly,
it is time to merge these settings back into the other config files.

The wave output frequency settings are updated to the coarser fre-
quency as that is more suitable for development. The prior (ops)
settings are noted as comments.

There are a couple of settings that control ice fields within FV3.
These are placed in `config.ice`, so when the ice model is on those
settings are still used.

Changes the output frequency from 3 to 6. While we may leave this
setting as the default for development, it is required right now
because the ocean frequency is hard-coded to 6 but links are created
based on `$FHOUT_GFS`. Resolving #1629 should correct this limitation.

"Gaussian" archive is updated to the former override values.

Resolves #1606
Refs #1629