Use deterministic marine resolutions for ensemble (#1894)
This PR determines the ensemble resolution for ocean, ice and waves based on the
deterministic configuration as there is no dual-resolution capability requested for
these model components.

For cycled DA, if the deterministic uses a coupled model, the ensemble will also be
a coupled model.  This can be turned off manually via adding `DO_AERO_ENKF`,
`DO_OCN_ENKF`, `DO_ICE_ENKF` and `DO_WAVE_ENKF` flags set in `config.base` or
`config.efcs`.  This is currently handled manually and will be automated when a
requirement arises for this flexibility.

Fixes #1730