Split clean-up into separate job
Moves the clean-up that was previously done in the archive jobs into
their own separate job. It is mostly a copy of what existed in the
archive scripts, except some adjustments to account for ensemble
members in the same job as the deterministic and some others for
shellcheck compliance/other standard code improvements.

Some settings used by both jobs were elevated to `config.base`.
Others only needed for cleanup were moved to the new config for that
job.

Resolves #583