Add stage_ic to GEFS job mesh and update dependencies
The structure of staged ICs have been updated to be more stream-
lined and also to allow ensemble members to support GEFS. The data
dependencies for the job are updated to use the new file structure.

The stage_ic job is also added to the GEFS job mesh to copy staged
initial conditions.

Refs #911