Streamline cleanup script
The old cleanup script was more convoluted than it needed to be.
The script has been streamlined quite a bit and now works properly.
Instead of going through every COM template, `COM_TOP is used along
with find to get everything at once (well, a few passes of find are
needed to deal with regular files, symlinks, and then directories).

Refs #583
Resolves #1872