Enable different coupled settings for ensemble
Preliminarily adds the ability to turn off some coupled components for
the ensemble.

To enable this, many settings are moved out of `config.fcst` and into
`config.ufs` so they are calculated after any overrides of `DO_*` have
been made. Some settings also had to be moved out of `config.base` for
the same reason.

The name of the nems configure template used is converted from a
partial file name to the full file name. In anticipation of #1942, the
filename is assigned to a variable named `ufs_configure_template`,
though otherwise 'nems' is still used for filenames, etc.

Refs #1692