Parallelize the build scripts (#1998)
Optimized/parallelized build scripts.

A few notes:

1) The default number of build jobs is 20, but one could argue for using 40.  When researching
this, I looked up what the SRW is compiling with.  That system uses 40 cores, which seems a
little excessive, but on testing the global workflow, the actual number of cores being used at
any given time rarely exceeds 16 when running with 40 cores.  This is because the builds tend to
use multiple threads in the beginning when compiling low-level modules while the higher-level
modules are more or less serial AND because the GDASApp takes several minutes to initialize
all of its subrepositories by which time the smaller builds are complete.

2) I also updated checkout.sh so that all checkouts are simultaneous.  The CPU load for
`git submodule` is quite low, so running 16 instead of 8 jobs at once is not much more expensive.

3) To make this work, I had to add `-j` options to most of the build scripts.  The only exception is
build_upp, for which the build script within the UPP is hard coded to use 6 cores.

4) I fixed a few small bugs in the build scripts along the way.

5) Lastly, this reduce the total build time from ~2.5 hours for the entire system (including GDAS
and GSI in the same build) to ~40 minutes when running with `-j 40`.

Resolves #1978