Update archive job dependencies for GSI monitor jobs
- Since the GSI monitor jobs don't run in the first half cycle
have updated the dependencies for the gdasarch job to add
additional logic for the first half cycle.
- The gdasarch job will now check a new "or" logic if the
GSI monitor job(s) are done or there isn't a prior cycle.
- The first half cycle gdasarch job should fire off
based on the lack of prior cycle instead of waiting for the
GSI monitor jobs that won't exist. Later cycles should wait
for the GSI monitor job(s) are complete.
- Also updated the dependencies to appropriately add the
gfsvminmon job for the gfsarch job when in cycled mode.

Refs #1908