Hack UPP version
The stack UPP was using on Orion was recently deleted, leaving us
unable to build the UFS verison of UPP there. UPP has an updated
version using spack-stack that works, so the checkout script is
modified to check out that version of UPP after UFS has been
checked out.

This temporary hack can be removed once we move to a UFS version
that advances the UPP version to or beyond 78f369b.

Resolves #2041