Restore undefined var error checking after module-setup
`module-setup.sh` was turning on error checking, but not turning it off
again at the end. Now the script will run `set_strict` if it is defined
(as during a workflow job) and otherwise turns off undefined variable
checking at the end. This will stop `gw_setup.sh` from breaking tab-
completion in an interactive shell.