Fix analysis and post on Orion (#2046)
Fixes a couple things to get workflow to work on Orion again:

First, while #2042 allowed post to be built, the `upp.x` executable would fail due
to a missing library. Now post jobs (temporarily) uses the UPP modules (plus
others needed for runtime) instead of the normal workflow runtime module. This
should actually fix #2041. This hack can likely be removed once we move to
spack-stack.

Second, the launcher command for the analysis jobs on Orion are updated to
specify the number of CPUs per task. This was necessary as a recent slurm update
on Orion discontinued reading that value from the scheduler. Other jobs will be
handled later (see #2044).

Additionally, a change is made to `module-setup.sh` that will eliminate the
annoyance of `gw_setup.sh` breaking tab completion because undefined variable
checking was turned on for the script but not back off.

Resolves #1996
Resolves #2041
Refs #2044