Refine fit2obs job in rocoto mesh (#2047)
This PR adjusts the `fit2obs` job in the rocoto mesh to better define when it exists
and how the archive job dependencies accommodate it not being in the first half
cycle.

Changes in PR:

1. Remove the `fit2obs` job from the first half cycle `cycledef`.
2. Adjust the archive job dependencies to now include the `fit2obs` job in the new
logic check added in PR #1983. The `gdasarch` job in the first half cycle will check if
`gdasfit2obs` is complete OR no prior cycle exists. Also cleaned up this new logic
check block some to make it a tad simpler.
3. Also, fix the addition of the `fit2obs` config to the config list in setup scripts. Now
tied to whether `fit2obs` job is on instead of always adding the config to the list.

Resolves #2043