Separate tracker and genesis jobs from vrfy job (#2058)
This PR creates new individual jobs for the tracker and genesis jobs that are being
pulled out of the `vrfy` job (issue #235). The `tracker` and `genesis` jobs will be on
by default while the `genesis_fsu` job will be off by default. This mimics the current
settings in the `vrfy` job. 

No enhancements or upgrades are made to the jobs themselves in this PR, only to
separate them from the `vrfy` job. However, one change compared to how they
currently run in the `vrfy` job is to no longer run the `tracker` in the gdas suite. This
mimics how the `ens_tracker` package is run in operations (gfs only).

Changes in this PR:

1. Create rocoto job scripts for tracker/genesis jobs.
2. Create configs for tracker/genesis jobs, as well as `config.tropcy` for shared
settings.
3. Add tracker/genesis jobs to `config.resources`; use resources from the ecf scripts
for `ens_tracker` jobs in operations
4. Add tracker/genesis jobs to rocoto setup for both cycled and forecast-only
modes.
5. Update arguments for `jjob_header.sh` script in tracker/genesis JJOB scripts to use
specific job names and configs instead of vrfy.
6. Rename `tracker_ver` to `ens_tracker_ver` in version files to mimic variable within
tracker package.
7. Rename and move tracker/genesis switches to `config.base`.
8. Remove tracker/genesis jobs from `vrfy` job (`jobs/rocoto/vrfy` and `config.vrfy`).
9. Reduce vrfy job resources to accommodate tasks having been removed.

Refs #235
Resolves #1988
Resolves #2048