Fix check for existence of preamble functions
A few of the setup scripts were checking for the existence of
functions set by the preamble before calling them so they could
also be used stand-alone. However, the `type` command returns `1`
if the function name does not exist, which will cause the script
to die is `set -e` is on. Now the `type` command fails over to an
empty string.