Add --cpus-per-task to srun commands for Hera and Orion (#2077)
This PR adds the `--cpus-per-task` flag to non-CFP `APRUN`/srun commands
in the env files for Orion and Hera. This is needed as a result of an upgrade
to SLURM on Orion that no longer exported the thread value to the
environment. Issue #1996 was a hotfix to add this flag for the analysis job
that was suddenly running slowly after the SLURM upgrade. A similar SLURM
upgrade was done on Hera but the sysadmins there implemented a
workaround so this issue did not appear there.

This PR adds the `--cpus-per-task` flag for both machines...needed on Orion
and as a safety measure on Hera.

Also included in this PR is a fix to indentations in the
`scripts/exglobal_diag.sh` script that was noticed during work for this PR.

Timing differences were not observed for other jobs after adding this srun
flag. Job timings were of a similar variation between control and test runs.
Any potential additional speedup was not observed. The test runs also
reproduced the control runs with `develop`.

Resolves #2044